requirements.txt
fastapi
uvicorn
sentry-sdk
python-decouple
redis
pika
pymongo